+++ CHANGELOG.md
@@ -25,6 +25,7 @@ This file should be updated before a new release is deployed.
 * Upgraded to rails 6.0 and ruby 2.6.5 [#214](https://gitlab.com/yale-sdmp/cardswipr/issues/214)
 * Updated event creation to allow events with preregistation [#220](https://gitlab.com/yale-sdmp/cardswipr/-/issues/220)
 * Updated attendance creator to handle events with preregistration [#221](https://gitlab.com/yale-sdmp/cardswipr/issues/221)
+* Changed Fake Person to return a person with the same netid that was passed to it [#232](https://gitlab.com/yale-sdmp/cardswipr/issues/232)
 
 ### Added
 * Added User model [#107](https://gitlab.com/yale-sdmp/cardswipr/issues/107)